+++ fixture_src/ConfigurationWithArrayEnum/FooEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Cspray\AnnotatedContainerFixture\ConfigurationWithArrayEnum;
+
+enum FooEnum {
+    case Bar;
+    case Baz;
+    case Qux;
+}